+++ server/src/main/java/com/codecool/controller/EventController.java
@@ -2,7 +2,6 @@ package com.codecool.controller;
 
 import com.codecool.DTO.event.EventDTO;
 import com.codecool.DTO.event.NewEventDTO;
-
 import com.codecool.DTO.tag.TagDTO;
 import com.codecool.service.EventService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,45 +14,44 @@ import java.util.List;
 
 public class EventController {
     private final EventService eventService;
-   @Autowired
+
+    @Autowired
     public EventController(EventService eventService) {
         this.eventService = eventService;
     }
 
     @GetMapping("/{eventId}")
     public EventDTO getEventById(@PathVariable int eventId) {
-       return eventService.getEventById(eventId);
+        return eventService.getEventById(eventId);
     }
 
     @PostMapping
-    public int createEvent(@RequestBody NewEventDTO newEventDTO) {
-       return eventService.addEvent(newEventDTO);
+    public long createEvent(@RequestBody NewEventDTO newEventDTO) {
+        return eventService.addEvent(newEventDTO);
     }
 
     @GetMapping
     public List<EventDTO> getAllEvents() {
-
-
-       return eventService.getAllEvents();
+        return eventService.getAllEvents();
     }
 
     @PatchMapping("/{eventId}/modify")
     public boolean modifyEvent(@PathVariable int eventId, @RequestBody EventDTO eventDTO) {
-       return eventService.modifyEvent(eventDTO);
+        return eventService.modifyEvent(eventDTO);
     }
 
     @DeleteMapping("/{eventId}")
-    public boolean deleteEvent(@PathVariable int eventId) {
-       return eventService.deleteEventById(eventId);
+    public boolean deleteEvent(@PathVariable long eventId) {
+        return eventService.deleteEventById(eventId);
     }
 
-@PostMapping("/{eventId}")
-    public boolean addEvent(@PathVariable int eventId, @RequestBody TagDTO tagDTO) {
-       return eventService.addTagToEvent(eventId, tagDTO);
-}
+    @PostMapping("/{eventId}")
+    public boolean addEvent(@PathVariable long eventId, @RequestBody TagDTO tagDTO) {
+        return eventService.addTagToEvent(eventId, tagDTO);
+    }
 
-@DeleteMapping("/tag/{eventId}")
-    public boolean deleteTag(@PathVariable int eventId, @RequestParam int tagId) {
-       return eventService.deleteTagFromEvent(eventId,tagId);
-}
+    @DeleteMapping("/tag/{eventId}")
+    public boolean deleteTag(@PathVariable long eventId, @RequestParam int tagId) {
+        return eventService.deleteTagFromEvent(eventId, tagId);
+    }
 }
